Merged PR 2219: Intern feature improvement
Adjust intern feature after person remodeling.
When all interns are requested the answer should contain "display name". The source of which is Azure AD if intern have work email and If doesn't have one it will be formed from first name and last name of that person. The email will be the one from endava if it is provided, if not - the personal email.
When there is a request for a specific intern, the response will contain all available information for him. In that case personal email is mandatory, but work email can be empty (for example if he has not been accepted for campaign). Model will provide first and last name, for "display name" same rule apply. The source  is Azure AD if intern have work email and If doesn't have one it will be formed from first name and last name of that person.
Updating first and last name will change "display name" only when intern is not with active status (doesn't have display name from azure).
Requesting all interns by campaign now return the same object as request for all interns (no matter of campaign).
When intern is accept to start in campaign and doesn't already have work email it should be added. Providing that email should trigger Azure AD invite and save it to database. That should change the return object structure, including display name of that person.

Related work items: #5077